[direct] move shiftedWhile to kyo-direct
