+++ kyo-direct/shared/src/main/scala/kyo/Direct.scala
@@ -4,11 +4,14 @@ import cps.CpsMonad
 import cps.CpsMonadContext
 import cps.async
 import cps.runtime.SeqAsyncShift
-import directInternal.*
+import directInternal.KyoCpsMonad
 import kyo.Ansi.*
+import kyo.Kyo.toIndexed
+import kyo.kernel.Loop
 import kyo.kernel.internal.Safepoint
 import scala.annotation.tailrec
 import scala.collection.SeqOps
+import scala.collection.immutable.LinearSeq
 import scala.quoted.*
 
 /** Defers the execution of a block of code, allowing the use of `.now` and `.later` for effect handling.
@@ -246,7 +249,7 @@ object directInternal:
             monad: CpsMonad[F]
         )(prolog: S, condition: A => F[Boolean], acc: (S, Boolean, A) => S, epilog: S => R): F[R] =
             monad match
-                case _: KyoCpsMonad[?] if useLoop(c) => Kyo.shiftedWhile(c)(prolog, condition, acc, epilog)
+                case _: KyoCpsMonad[?] if useLoop(c) => directInternal.shiftedWhile(c)(prolog, condition, acc, epilog)
                 case _                               => super.shiftedWhile(c, monad)(prolog, condition, acc, epilog)
 
         override def map[F[_], B](c: CA, monad: CpsMonad[F])(f: A => F[B]): F[C[B]] =
@@ -265,4 +268,40 @@ object directInternal:
 
     transparent inline given shiftedSeqToChunk[A]: SeqToChunkAsyncShift[A] = new SeqToChunkAsyncShift[A]
     transparent inline given shiftedChunk[A]: ChunkAsyncShift[A]           = new ChunkAsyncShift[A]
+
+    /** from Dotty-Cps-Async override def shiftedWhile[F[_], S, R]( c: CA, monad: CpsMonad[F] )(prolog: S, condition: A => F[Boolean], acc:
+      * (S, Boolean, A) => S, epilog: S => R): F[R]
+      */
+    private[kyo] def shiftedWhile[A, S, B, C](source: IterableOnce[A])(
+        prolog: B,
+        f: Safepoint ?=> A => Boolean < S,
+        acc: (B, Boolean, A) => B,
+        epilog: B => C
+    )(using Frame, Safepoint): C < S =
+        source.knownSize match
+            case 0 => epilog(prolog)
+            case _ =>
+                source match
+                    case linearSeq: LinearSeq[A] => Loop(linearSeq, prolog): (seq, b) =>
+                            if seq.isEmpty then Loop.done(epilog(b))
+                            else
+                                val curr = seq.head
+                                f(curr).map:
+                                    case true  => Loop.continue(seq.tail, acc(b, true, curr))
+                                    case false => Loop.done(epilog(acc(b, false, curr)))
+
+                    case other =>
+                        val indexedSeq = toIndexed(other)
+                        val size       = indexedSeq.length
+                        Loop.indexed(prolog): (idx, b) =>
+                            if idx == size then Loop.done(epilog(b))
+                            else
+                                val curr = indexedSeq(idx)
+                                f(curr).map:
+                                    case true  => Loop.continue(acc(b, true, curr))
+                                    case false => Loop.done(epilog(acc(b, false, curr)))
+
+        end match
+    end shiftedWhile
+
 end directInternal
