+++ kyo-kernel/shared/src/main/scala/kyo/Kyo.scala
@@ -454,41 +454,6 @@ object Kyo:
                                     case false => Loop.done(acc)
     end takeWhile
 
-    /** from Dotty-Cps-Async override def shiftedWhile[F[_], S, R]( c: CA, monad: CpsMonad[F] )(prolog: S, condition: A => F[Boolean], acc:
-      * (S, Boolean, A) => S, epilog: S => R): F[R]
-      */
-    private[kyo] def shiftedWhile[A, S, B, C](source: IterableOnce[A])(
-        prolog: B,
-        f: Safepoint ?=> A => Boolean < S,
-        acc: (B, Boolean, A) => B,
-        epilog: B => C
-    )(using Frame, Safepoint): C < S =
-        source.knownSize match
-            case 0 => epilog(prolog)
-            case _ =>
-                source match
-                    case linearSeq: LinearSeq[A] => Loop(linearSeq, prolog): (seq, b) =>
-                            if seq.isEmpty then Loop.done(epilog(b))
-                            else
-                                val curr = seq.head
-                                f(curr).map:
-                                    case true  => Loop.continue(seq.tail, acc(b, true, curr))
-                                    case false => Loop.done(epilog(acc(b, false, curr)))
-
-                    case other =>
-                        val indexedSeq = toIndexed(other)
-                        val size       = indexedSeq.length
-                        Loop.indexed(prolog): (idx, b) =>
-                            if idx == size then Loop.done(epilog(b))
-                            else
-                                val curr = indexedSeq(idx)
-                                f(curr).map:
-                                    case true  => Loop.continue(acc(b, true, curr))
-                                    case false => Loop.done(epilog(acc(b, false, curr)))
-
-        end match
-    end shiftedWhile
-
     /** Drops elements from an `IterableOnce` while a predicate holds true.
       *
       * @param source
@@ -539,7 +504,7 @@ object Kyo:
             else v.map(t => Loop.continue(acc.append(t)))
         }
 
-    private def toIndexed[A](source: IterableOnce[A]): Seq[A] =
+    private[kyo] def toIndexed[A](source: IterableOnce[A]): Seq[A] =
         source match
             case seq: IndexedSeq[A] => seq
             case other              => Chunk.from(other)
